requirements.txt
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
google-api-python-client>=2.0.0
anthropic>=0.25.0
streamlit>=1.30.0
plotly>=5.0.0
pandas>=2.0.0
